refactor: remove stream classes from privmx-endpoint-streams and use those from privmx-endpoint
