--- privmx-endpoint-streams/android/src/main/java/com/simplito/java/privmx_endpoint/model/PublishedStreamData.java
@@ -1,13 +1,0 @@
-package com.simplito.java.privmx_endpoint.model;
-
-public class PublishedStreamData {
-    public String streamRoomId;
-    public StreamInfo stream;
-    public String userId;
-
-    public PublishedStreamData(String streamRoomId, StreamInfo stream, String userId) {
-        this.streamRoomId = streamRoomId;
-        this.stream = stream;
-        this.userId = userId;
-    }
-}
