--- privmx-endpoint-streams/android/src/main/java/com/simplito/java/privmx_endpoint/model/Settings.java
@@ -1,6 +1,0 @@
-package com.simplito.java.privmx_endpoint.model;
-
-public class Settings {
-    public Settings() {
-    }
-}
