--- privmx-endpoint-streams/android/src/main/java/com/simplito/java/privmx_endpoint/model/StreamTrackModificationPair.java
@@ -1,19 +1,0 @@
-package com.simplito.java.privmx_endpoint.model;
-
-public class StreamTrackModificationPair {
-    public StreamTrackInfo before;
-    public StreamTrackInfo after;
-
-    public StreamTrackModificationPair(StreamTrackInfo before, StreamTrackInfo after) {
-        this.before = before;
-        this.after = after;
-    }
-
-    public StreamTrackModificationPair(StreamTrackInfo before) {
-        this(before, null);
-    }
-
-    public StreamTrackModificationPair() {
-        this(null, null);
-    }
-}
