+++ privmx-endpoint-streams/android/src/main/java/com/simplito/java/privmx_endpoint/modules/stream/StreamApi.java
@@ -14,16 +14,15 @@ import com.simplito.java.privmx_endpoint.model.ContainerPolicy;
 import com.simplito.java.privmx_endpoint.model.DeviceType;
 import com.simplito.java.privmx_endpoint.model.MediaDevice;
 import com.simplito.java.privmx_endpoint.model.PagingList;
-import com.simplito.java.privmx_endpoint.model.Settings;
-import com.simplito.java.privmx_endpoint.model.StreamHandle;
-import com.simplito.java.privmx_endpoint.model.StreamInfo;
-import com.simplito.java.privmx_endpoint.model.StreamPublishResult;
-import com.simplito.java.privmx_endpoint.model.StreamRoom;
-import com.simplito.java.privmx_endpoint.model.StreamSettings;
-import com.simplito.java.privmx_endpoint.model.StreamSubscription;
 import com.simplito.java.privmx_endpoint.model.UserWithPubKey;
 import com.simplito.java.privmx_endpoint.model.events.eventSelectorTypes.StreamEventSelectorType;
 import com.simplito.java.privmx_endpoint.model.events.eventTypes.StreamEventType;
+import com.simplito.java.privmx_endpoint.model.stream.Settings;
+import com.simplito.java.privmx_endpoint.model.stream.StreamInfo;
+import com.simplito.java.privmx_endpoint.model.stream.StreamPublishResult;
+import com.simplito.java.privmx_endpoint.model.stream.StreamRoom;
+import com.simplito.java.privmx_endpoint.model.stream.StreamHandle;
+import com.simplito.java.privmx_endpoint.model.stream.StreamSubscription;
 
 import org.webrtc.AudioSource;
 import org.webrtc.AudioTrack;
