--- privmx-endpoint-streams/jvm/src/main/java/com/simplito/java/privmx_endpoint/model/KeyType.java
@@ -1,17 +1,0 @@
-//
-// PrivMX Endpoint Java.
-// Copyright © 2024 Simplito sp. z o.o.
-//
-// This file is part of the PrivMX Platform (https://privmx.dev).
-// This software is Licensed under the MIT License.
-//
-// See the License for the specific language governing permissions and
-// limitations under the License.
-//
-
-package com.simplito.java.privmx_endpoint.model;
-
-public enum KeyType {
-    LOCAL,
-    REMOTE
-}
