--- privmx-endpoint-streams/jvm/src/main/java/com/simplito/java/privmx_endpoint/model/MediaDevice.java
@@ -1,13 +1,0 @@
-package com.simplito.java.privmx_endpoint.model;
-
-public class MediaDevice {
-    public String name;
-    public String id;
-    public DeviceType type;
-
-    public MediaDevice(String name, String id, DeviceType type) {
-        this.name = name;
-        this.id = id;
-        this.type = type;
-    }
-}
