--- privmx-endpoint-streams/jvm/src/main/java/com/simplito/java/privmx_endpoint/model/StreamHandle.java
@@ -1,13 +1,0 @@
-package com.simplito.java.privmx_endpoint.model;
-
-public class StreamHandle {
-    private final Long value;
-
-    public StreamHandle(Long value) {
-        this.value = value;
-    }
-
-    public Long getValue() {
-        return value;
-    }
-}
