--- privmx-endpoint-streams/jvm/src/main/java/com/simplito/java/privmx_endpoint/model/StreamSubscription.java
@@ -1,15 +1,0 @@
-package com.simplito.java.privmx_endpoint.model;
-
-public class StreamSubscription {
-    public long streamId;
-    public String streamTrackId;
-
-    public StreamSubscription(long streamId, String streamTrackId) {
-        this.streamId = streamId;
-        this.streamTrackId = streamTrackId;
-    }
-
-    public StreamSubscription(long streamId) {
-        this.streamId = streamId;
-    }
-}
